Move ConsoleAssertionPrinter to catch_reporter_bases
